requirements.txt
requests
schedule
pytz
python-dotenv
Pillow
Flask
waitress
redis
hiredis
